Bug 1433158 - Update libvpx to 1.7.0; r=jya
Reviewers: jya

Tags: #secure-revision

Bug #: 1433158

Differential Revision: https://phabricator.services.mozilla.com/D2940

UltraBlame original commit: 04a7c8e2ab6d9be7b4c083ed14104005e4da87b7

+++ media/libvpx/libvpx/vpx_dsp/x86/fwd_dct32x32_impl_avx2.h
@@ -51,7 +51,7 @@ void FDCT32x32_2D_AVX2(const int16_t *input, int16_t *output_org, int stride) {
   
   
   
-  const __m256i k__cospi_p16_p16 = _mm256_set1_epi16((int16_t)cospi_16_64);
+  const __m256i k__cospi_p16_p16 = _mm256_set1_epi16(cospi_16_64);
   const __m256i k__cospi_p16_m16 =
       pair256_set_epi16(+cospi_16_64, -cospi_16_64);
   const __m256i k__cospi_m08_p24 = pair256_set_epi16(-cospi_8_64, cospi_24_64);
